fix issue FSX Lustre not mounted in Downloader EC2 instance

+++ cdk_packages/assets/download_files.sh
@@ -1,48 +1,73 @@
 #!/bin/bash
 # log UserData script output, source: https://alestic.com/2010/12/ec2-user-data-output/
-exec > >(tee /var/log/user-data.log | logger -t user-data -s 2>/dev/console) 2>&1
+exec > >(tee -a /var/log/download-files.log | logger -t user-data -s 2>/dev/console) 2>&1
 echo -----
-echo ----- start UserData script
+echo ----- start script
 echo -----
 
 echo ----- update system -----
 
-apt-get update
-apt-get upgrade -y
-apt-get install unzip -y
-apt-get install jq -y
+INDICATOR=/var/tmp/indicator-update-system
+if [ ! -f "${INDICATOR}" ]; then
+    apt-get update
+    apt-get upgrade -y
+    touch "${INDICATOR}"
+fi
 
 echo ----- install CloudWatch agent -----
 
-wget https://s3.amazonaws.com/amazoncloudwatch-agent/ubuntu/arm64/latest/amazon-cloudwatch-agent.deb
-dpkg -i -E amazon-cloudwatch-agent.deb
-amazon-cloudwatch-agent-ctl -a fetch-config -m ec2 -s -c ssm:/ONT-performance-benchmark/downloader-cloudwatch-agent-config
-
-echo ----- install AWS CLI -----
-
-curl "https://awscli.amazonaws.com/awscli-exe-linux-aarch64.zip" -o "awscliv2.zip"
-unzip -qq awscliv2.zip
-./aws/install
+INDICATOR=/var/tmp/indicator-cloudwatch-agent
+if [ ! -f "${INDICATOR}" ]; then
+    wget https://s3.amazonaws.com/amazoncloudwatch-agent/ubuntu/arm64/latest/amazon-cloudwatch-agent.deb
+    #wget https://amazoncloudwatch-agent.s3.amazonaws.com/ubuntu/amd64/latest/amazon-cloudwatch-agent.deb
+    dpkg -i -E amazon-cloudwatch-agent.deb
+    amazon-cloudwatch-agent-ctl -a fetch-config -m ec2 -s -c ssm:/ONT-performance-benchmark/downloader-cloudwatch-agent-config
+    touch "${INDICATOR}"
+fi
 
 echo ----- reset status parameters -----
 
-aws ssm put-parameter --name /ONT-performance-benchmark/download-status --value "not started" --overwrite --output text
-aws ssm put-parameter --name /ONT-performance-benchmark/pod5-converter-status --value "not started" --overwrite --output text
+INDICATOR=/var/tmp/indicator-reset-status-parameters
+if [ ! -f "${INDICATOR}" ]; then
+    aws ssm put-parameter --name /ONT-performance-benchmark/download-status --value "not started" --overwrite --output text
+    aws ssm put-parameter --name /ONT-performance-benchmark/pod5-converter-status --value "not started" --overwrite --output text
+    touch "${INDICATOR}"
+fi
 
 echo ----- install FSx for Lustre client -----
 
-wget -O - https://fsx-lustre-client-repo-public-keys.s3.amazonaws.com/fsx-ubuntu-public-key.asc | gpg --dearmor | sudo tee /usr/share/keyrings/fsx-ubuntu-public-key.gpg >/dev/null
-bash -c 'echo "deb [signed-by=/usr/share/keyrings/fsx-ubuntu-public-key.gpg] https://fsx-lustre-client-repo.s3.amazonaws.com/ubuntu jammy main" > /etc/apt/sources.list.d/fsxlustreclientrepo.list && apt-get update'
-apt-get install -y linux-image-5.15.0-1039-aws
-sed -i 's/GRUB_DEFAULT=.\+/GRUB\_DEFAULT="Advanced options for Ubuntu>Ubuntu, with Linux 5.15.0-1039-aws"/' /etc/default/grub
-update-grub
-apt-get install -y linux-aws lustre-client-modules-$(uname -r)
+INDICATOR=/var/tmp/indicator-install-kernel
+if [ ! -f "${INDICATOR}" ]; then
+    wget -O - https://fsx-lustre-client-repo-public-keys.s3.amazonaws.com/fsx-ubuntu-public-key.asc | gpg --dearmor | sudo tee /usr/share/keyrings/fsx-ubuntu-public-key.gpg >/dev/null
+    bash -c 'echo "deb [signed-by=/usr/share/keyrings/fsx-ubuntu-public-key.gpg] https://fsx-lustre-client-repo.s3.amazonaws.com/ubuntu jammy main" > /etc/apt/sources.list.d/fsxlustreclientrepo.list && apt-get update'
+    version="6.2.0-1018-aws"
+    apt-get install -y linux-image-"$version"
+    sed -i 's/GRUB_DEFAULT=.\+/GRUB\_DEFAULT="Advanced options for Ubuntu>Ubuntu, with Linux '$version'"/' /etc/default/grub
+    update-grub
+    touch "${INDICATOR}"
+    echo "Reboot to activate installed kernel."
+    shutdown -r now
+    sleep 60
+fi
 
-echo ----- install python and libraries -----
+INDICATOR=/var/tmp/indicator-install-fsx-lustre
+if [ ! -f "${INDICATOR}" ]; then
+    apt-get install -y linux-aws lustre-client-modules-"$(uname -r)"
+    touch "${INDICATOR}"
+    echo "Reboot to activate FSX Lustre client."
+    shutdown -r now
+    sleep 60
+fi
 
-apt-get install python3-pip -y
-pip install pod5
-pip install pandas
+echo ----- install Pod5 -----
+
+INDICATOR=/var/tmp/indicator-install-pod5
+if [ ! -f "${INDICATOR}" ]; then
+    apt-get install python3-pip -y
+    pip install pod5
+    pip install pandas
+    touch "${INDICATOR}"
+fi
 
 echo ----- get token to access instance metadata -----
 
@@ -58,36 +83,48 @@ mount -t lustre -o noatime,flock "${fsx_dns_mount_name}" /fsx
 
 echo ----- download test data -----
 
-## DISCLAIMER: The download URL is from the CliveOME 5mC dataset (ONLA29134 size: 745.4 GiB). A FAST5 data set published by
-## Oxford Nanopore Technologies. For more details about the data set, please see
-## https://labs.epi2me.io/cliveome_5mc_cfdna_celldna/
-download_url='s3://ont-open-data/cliveome_kit14_2022.05/gdna/flowcells/ONLA29134/20220510_1127_5H_PAM63974_a5e7a202/fast5_pass/'
+INDICATOR=/var/tmp/indicator-download-test-data
+if [ ! -f "${INDICATOR}" ]; then
+    # DISCLAIMER: The download URL is from the CliveOME 5mC dataset (ONLA29134 size: 745.4 GiB). A FAST5 data set published by
+    # Oxford Nanopore Technologies. For more details about the data set, please see
+    # https://labs.epi2me.io/cliveome_5mc_cfdna_celldna/
+    download_url='s3://ont-open-data/cliveome_kit14_2022.05/gdna/flowcells/ONLA29134/20220510_1127_5H_PAM63974_a5e7a202/fast5_pass/'
 
-files_to_download=($(aws s3 ls $download_url --no-sign-request | awk '{print $4}' | sort -n -t _ -k 4))
-local_s3_url=$(eval 'aws ssm get-parameters --names /ONT-performance-benchmark/data-s3-bucket --query '"'"'Parameters[0].Value'"'"' --output text')
-aws ssm put-parameter --name /ONT-performance-benchmark/download-status --value "in progress" --overwrite --output text
-for ((i = 0; i < ${#files_to_download[@]}; ++i)); do
-    current=$(( i + 1 ))
-    echo "$current/${#files_to_download[@]}: ${files_to_download[$i]}"
-    aws s3 cp "$download_url${files_to_download[$i]}" - --no-sign-request | aws s3 cp - "s3://$local_s3_url/fast5-all-files/${files_to_download[$i]}"
-done
-aws ssm put-parameter --name /ONT-performance-benchmark/download-status --value "completed" --overwrite --output text
+    files_to_download=($(aws s3 ls $download_url --no-sign-request | awk '{print $4}' | sort -n -t _ -k 4))
+    local_s3_url=$(eval 'aws ssm get-parameters --names /ONT-performance-benchmark/data-s3-bucket --query '"'"'Parameters[0].Value'"'"' --output text')
+    aws ssm put-parameter --name /ONT-performance-benchmark/download-status --value "in progress" --overwrite --output text
+    for ((i = 0; i < ${#files_to_download[@]}; ++i)); do
+        current=$(( i + 1 ))
+        echo "$current/${#files_to_download[@]}: ${files_to_download[$i]}"
+        aws s3 cp "$download_url${files_to_download[$i]}" - --no-sign-request | aws s3 cp - "s3://$local_s3_url/fast5-all-files/${files_to_download[$i]}"
+    done
+    aws ssm put-parameter --name /ONT-performance-benchmark/download-status --value "completed" --overwrite --output text
+    touch "${INDICATOR}"
+fi
 
 echo ----- convert FAST5 to POD5 -----
 
-## For details about how top convert FAST5 to POD5 please see
-## https://github.com/nanoporetech/pod5-file-format/blob/master/python/pod5/README.md
-aws ssm put-parameter --name /ONT-performance-benchmark/pod5-converter-status --value "in progress" --overwrite --output text
-num_fast5_files=$(find /fsx/fast5-all-files -name "*.fast5" | wc -l)
-echo "Total number of fast5 files: ${num_fast5_files}"
-pod5 convert fast5 /fsx/fast5-all-files/*.fast5 --output /fsx/pod5-all-files/ --one-to-one /fsx/fast5-all-files/ --threads 20 --strict --force-overwrite
-aws ssm put-parameter --name /ONT-performance-benchmark/pod5-converter-status --value "completed" --overwrite --output text
+INDICATOR=/var/tmp/indicator-convert-fast5-to-pod5
+if [ ! -f "${INDICATOR}" ]; then
+    ## For details about how top convert FAST5 to POD5 please see
+    ## https://github.com/nanoporetech/pod5-file-format/blob/master/python/pod5/README.md
+    aws ssm put-parameter --name /ONT-performance-benchmark/pod5-converter-status --value "in progress" --overwrite --output text
+    num_fast5_files=$(find /fsx/fast5-all-files -name "*.fast5" | wc -l)
+    echo "Total number of fast5 files: ${num_fast5_files}"
+    pod5 convert fast5 /fsx/fast5-all-files/*.fast5 --output /fsx/pod5-all-files/ --one-to-one /fsx/fast5-all-files/ --threads 20 --strict --force-overwrite
+    aws ssm put-parameter --name /ONT-performance-benchmark/pod5-converter-status --value "completed" --overwrite --output text
+    touch "${INDICATOR}"
+fi
 
 echo ----- create test data sets -----
 
-script_file=$(eval '/usr/local/bin/aws ssm get-parameters --region '"$region"' --names /ONT-performance-benchmark/pod5-create-test-data-script --query '"'"'Parameters[0].Value'"'"' --output text')
-aws s3 cp "$script_file" create_test_data_sets.py
-python3 create_test_data_sets.py
+INDICATOR=/var/tmp/indicator-create-test-data-sets
+if [ ! -f "${INDICATOR}" ]; then
+    script_file=$(eval '/usr/local/bin/aws ssm get-parameters --region '"$region"' --names /ONT-performance-benchmark/pod5-create-test-data-script --query '"'"'Parameters[0].Value'"'"' --output text')
+    aws s3 cp "$script_file" create_test_data_sets.py
+    python3 create_test_data_sets.py
+    touch "${INDICATOR}"
+fi
 
 echo ----- check download and conversion results -----
